+++ src/ts/behaviours/Rotate.ts
@@ -1,4 +1,4 @@
-import Behaviour from "./Behaviour";
+import Behaviour from "../Behaviour";
 
 interface RotateParameters {
     speed : number
